+++ socialbot/__init__.py
@@ -4,6 +4,6 @@ Post, schedule, automate and analyze 13+ social networks from one place.
 """
 
 __app_name__ = "socialbot"
-__version__ = "1.0.0"
+__version__ = "1.0.1"
 
 from .models import Post, PublishResult, PostStatus  # noqa: F401